feat:feature:scheduler TagsScreen tagList and customTagList

+++ feature/scheduler/src/main/java/fit/asta/health/feature/scheduler/ui/viewmodel/SchedulerViewModel.kt
@@ -4,7 +4,9 @@ import android.content.Context
 import android.media.RingtoneManager
 import android.net.Uri
 import android.util.Log
+import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
+import androidx.core.net.toUri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -15,7 +17,9 @@ import fit.asta.health.common.utils.getCurrentTime
 import fit.asta.health.data.scheduler.db.entity.AlarmEntity
 import fit.asta.health.data.scheduler.db.entity.TagEntity
 import fit.asta.health.data.scheduler.remote.net.scheduler.Time
-import fit.asta.health.data.scheduler.remote.net.tag.ScheduleTagNetData
+import fit.asta.health.data.scheduler.remote.net.tag.Data
+import fit.asta.health.data.scheduler.remote.net.tag.NetCustomTag
+import fit.asta.health.data.scheduler.remote.toTagEntity
 import fit.asta.health.data.scheduler.repo.AlarmBackendRepo
 import fit.asta.health.data.scheduler.repo.AlarmLocalRepo
 import fit.asta.health.datastore.PrefManager
@@ -27,17 +31,12 @@ import fit.asta.health.feature.scheduler.ui.screen.alarmsetingscreen.AdvUiState
 import fit.asta.health.feature.scheduler.ui.screen.alarmsetingscreen.IvlUiState
 import fit.asta.health.feature.scheduler.ui.screen.alarmsetingscreen.TimeUi
 import fit.asta.health.feature.scheduler.ui.screen.alarmsetingscreen.ToneUiState
-import fit.asta.health.feature.scheduler.ui.screen.tagscreen.TagState
-import fit.asta.health.feature.scheduler.ui.screen.tagscreen.TagsUiState
 import fit.asta.health.feature.scheduler.util.StateManager
 import fit.asta.health.feature.scheduler.util.VibrationPattern
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import java.text.DateFormat
 import java.text.SimpleDateFormat
@@ -66,15 +65,12 @@ class SchedulerViewModel
     val timeSettingUiState = interval.asStateFlow()
     private val advancedReminder = mutableStateOf(AdvUiState())
 
-    private val _tagsUiState = MutableStateFlow(TagsUiState())
-    val tagsUiState = _tagsUiState.asStateFlow()
+    private val _tagsList = mutableStateListOf<TagEntity>()
+    val tagsList = MutableStateFlow(_tagsList)
+    private val _customTagList = mutableStateListOf<TagEntity>()
+    val customTagList = MutableStateFlow(_customTagList)
 
 //    private val userId = "6309a9379af54f142c65fbfe"
-
-    val areInputsValid =
-        combine(alarmSettingUiState, tagsUiState) { alarm, _ ->
-            alarm.alarmName.isNotEmpty() && alarm.alarmDescription.isNotEmpty() && alarm.tagName.isNotEmpty()
-        }.stateIn(viewModelScope, SharingStarted.WhileSubscribed(1000), false)
 
     init {
         getEditUiData()
@@ -127,7 +123,6 @@ class SchedulerViewModel
                     VibrationPattern.Short -> "Short"
                     VibrationPattern.Long -> "Long"
                     VibrationPattern.Intermittent -> "Intermittent"
-                    else -> "Short"
                 }
             )
     }
@@ -174,7 +169,15 @@ class SchedulerViewModel
     fun getTagData() {
         viewModelScope.launch {
             alarmLocalRepo.getAllTags().collect {
-                _tagsUiState.value = _tagsUiState.value.copy(tagsList = it)
+                _tagsList.clear()
+                _customTagList.clear()
+                it.forEach { tagEntity ->
+                    if (tagEntity.uid == uId) {
+                        _customTagList.add(tagEntity)
+                    } else {
+                        _tagsList.add(tagEntity)
+                    }
+                }
                 Log.d("manish", "getTagData: $it")
             }
         }
@@ -184,8 +187,8 @@ class SchedulerViewModel
     fun updateServerTag(label: String, url: String) {//change
         viewModelScope.launch {
             val result = backendRepo.updateScheduleTag(
-                schedule = ScheduleTagNetData(
-                    uid = uId, type = 1, tag = label, url = url
+                schedule = NetCustomTag(
+                    uid = uId, name = label, localUrl = url.toUri()
                 )
             )
             when (result) {
@@ -201,29 +204,46 @@ class SchedulerViewModel
 
 
     fun selectedTag(tag: TagEntity) {
-        tag.selected = !tag.selected
-        _tagsUiState.value = _tagsUiState.value.copy(
-            selectedTag = TagState(
-                selected = tag.selected, meta = tag.meta, id = tag.id
-            )
-        )
         _alarmSettingUiState.value = _alarmSettingUiState.value.copy(
-            tagId = tag.meta.id, tagName = tag.meta.name, tagUrl = tag.meta.url
+            tagId = tag.id, tagName = tag.name, tagUrl = tag.url
         )
     }
 
-    fun undoTag(tag: TagEntity) {
-        viewModelScope.launch {
-            alarmLocalRepo.insertTag(tag)
-        }
-    }
 
     fun deleteTag(tag: TagEntity) {
         viewModelScope.launch {
-            alarmLocalRepo.deleteTag(tag)
+            when (backendRepo.deleteTagFromBackend(tag.uid, tag.id)) {
+                is ResponseState.Success -> {
+                    alarmLocalRepo.deleteTag(tag)
+                }
+
+                else -> {}
+            }
         }
     }
 
+    fun getTagDataFromServer() {
+        viewModelScope.launch {
+            when (val result = backendRepo.getTagListFromBackend(uId)) {
+                is ResponseState.Success -> {
+                    result.data.let { schedulerGetTagsList ->
+                        schedulerGetTagsList.list.forEach { tag ->
+                            insertTag(tag)
+                        }
+                        schedulerGetTagsList.customTagList?.forEach { tag ->
+                            insertTag(tag)
+                        }
+                    }
+                }
+
+                else -> {}
+            }
+        }
+    }
+
+    private suspend fun insertTag(tag: Data) {
+        alarmLocalRepo.insertTag(tag.toTagEntity())
+    }
 
     fun setSnoozeTime(snoozeTime: Int) {
         interval.value = interval.value.copy(snoozeTime = snoozeTime)
